remove incorrect test
