+++ alerting/src/test/kotlin/org/opensearch/alerting/MonitorDataSourcesIT.kt
@@ -374,43 +374,6 @@ class MonitorDataSourcesIT : AlertingSingleNodeTestCase() {
         )
     }
 
-    fun `test execute GetFindingsAction with params monitorId and findingIndex`() {
-        val docQuery = DocLevelQuery(query = "test_field:\"us-west-2\"", name = "3")
-        val docLevelInput = DocLevelMonitorInput("description", listOf(index), listOf(docQuery))
-        val trigger = randomDocumentLevelTrigger(condition = ALWAYS_RUN)
-        val customFindingsIndex = "custom_findings_index"
-        var monitor = randomDocumentLevelMonitor(
-            inputs = listOf(docLevelInput),
-            triggers = listOf(trigger),
-            dataSources = DataSources(findingsIndex = customFindingsIndex)
-        )
-        val monitorResponse = createMonitor(monitor)
-        val testTime = DateTimeFormatter.ISO_OFFSET_DATE_TIME.format(ZonedDateTime.now().truncatedTo(MILLIS))
-        val testDoc = """{
-            "message" : "This is an error from IAD region",
-            "test_strict_date_time" : "$testTime",
-            "test_field" : "us-west-2"
-        }"""
-        assertFalse(monitorResponse?.id.isNullOrEmpty())
-        monitor = monitorResponse!!.monitor
-        indexDoc(index, "1", testDoc)
-        val monitorId = monitorResponse.id
-        val executeMonitorResponse = executeMonitor(monitor, monitorId, false)
-        Assert.assertEquals(executeMonitorResponse!!.monitorRunResult.monitorName, monitor.name)
-        Assert.assertEquals(executeMonitorResponse.monitorRunResult.triggerResults.size, 1)
-        searchAlerts(monitorId)
-        val findings = searchFindings(monitorId, customFindingsIndex)
-        assertEquals("Findings saved for test monitor", 1, findings.size)
-        assertTrue("Findings saved for test monitor", findings[0].relatedDocIds.contains("1"))
-        // fetch findings - pass both monitorId and findingIndexName name. Monitor id should be ignored
-        val findingsFromAPI = getFindings(findings.get(0).id, "incorrect_monitor_id", customFindingsIndex)
-        assertEquals(
-            "Findings mismatch between manually searched and fetched via GetFindingsAction",
-            findings.get(0).id,
-            findingsFromAPI.get(0).id
-        )
-    }
-
     fun `test execute GetFindingsAction with unknown monitorId`() {
         val docQuery = DocLevelQuery(query = "test_field:\"us-west-2\"", name = "3")
         val docLevelInput = DocLevelMonitorInput("description", listOf(index), listOf(docQuery))
@@ -491,5 +454,60 @@ class MonitorDataSourcesIT : AlertingSingleNodeTestCase() {
                 )
             }
         }
+    }
+
+    fun `test get alerts by list of monitors containing both existent and non-existent ids`() {
+        val docQuery = DocLevelQuery(query = "test_field:\"us-west-2\"", name = "3")
+        val docLevelInput = DocLevelMonitorInput("description", listOf(index), listOf(docQuery))
+        val trigger = randomDocumentLevelTrigger(condition = ALWAYS_RUN)
+        var monitor = randomDocumentLevelMonitor(
+            inputs = listOf(docLevelInput),
+            triggers = listOf(trigger),
+        )
+        val monitorResponse = createMonitor(monitor)
+        val testTime = DateTimeFormatter.ISO_OFFSET_DATE_TIME.format(ZonedDateTime.now().truncatedTo(MILLIS))
+        val testDoc = """{
+            "message" : "This is an error from IAD region",
+            "test_strict_date_time" : "$testTime",
+            "test_field" : "us-west-2"
+        }"""
+
+        monitor = monitorResponse!!.monitor
+
+        val id = monitorResponse.id
+
+        var monitor1 = randomDocumentLevelMonitor(
+            inputs = listOf(docLevelInput),
+            triggers = listOf(trigger),
+        )
+        val monitorResponse1 = createMonitor(monitor1)
+        monitor1 = monitorResponse1!!.monitor
+        val id1 = monitorResponse1.id
+        indexDoc(index, "1", testDoc)
+        executeMonitor(monitor1, id1, false)
+        executeMonitor(monitor, id, false)
+        val alerts = searchAlerts(id)
+        assertEquals("Alert saved for test monitor", 1, alerts.size)
+        val alerts1 = searchAlerts(id)
+        assertEquals("Alert saved for test monitor", 1, alerts1.size)
+        val table = Table("asc", "id", null, 1000, 0, "")
+        var getAlertsResponse = client()
+            .execute(
+                AlertingActions.GET_ALERTS_ACTION_TYPE,
+                GetAlertsRequest(table, "ALL", "ALL", null, null)
+            )
+            .get()
+
+        Assert.assertTrue(getAlertsResponse != null)
+        Assert.assertTrue(getAlertsResponse.alerts.size == 2)
+
+        var alertsResponseForRequestWithoutCustomIndex = client()
+            .execute(
+                AlertingActions.GET_ALERTS_ACTION_TYPE,
+                GetAlertsRequest(table, "ALL", "ALL", null, null, listOf(id, id1, "1", "2"))
+            )
+            .get()
+        Assert.assertTrue(alertsResponseForRequestWithoutCustomIndex != null)
+        Assert.assertTrue(alertsResponseForRequestWithoutCustomIndex.alerts.size == 2)
     }
 }
